+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Fixed
+
+- hotfix: ([115](https://github.com/meateam/drive-ui/pull/115)): fix edit file name extention and special chars in the file name.
+- hotfix: ([114](https://github.com/meateam/drive-ui/pull/114)): fix drag and drop non responding page.
+- hotfix: ([49](https://github.com/meateam/drive-ui/pull/49/)) : fix fetch shared files bug.
+
 ## [v3.9.1] - 2021-06-23
  
 ### Feature
@@ -104,7 +110,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - add popup for convert docs.
-
 - add pagination to shared files and external transferd.
 -